Move shim classes into shim headers
